Just need one .editorconfig file - move .editorconfig files to repo root - also move `SalesforceResultCreator` to `...Internal` namespace
nit: add a few trailing newlines to files
