+++ src/Microsoft.AspNetCore.WebHooks.Receivers.Kudu/KuduWebHookAttribute.cs
